+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Updated bundled `composer.phar` binary to latest version `2.0.9`
     * Improved session fixation handling in PHP 7.4+ (cannot fix it in PHP 7.3 due to PHP bug)
     * Added optional password/database attributes for redis in `system.yaml`
+    * Added ability to filter enabled or disabled with bin/gpm index [#3187](https://github.com/getgrav/grav/pull/3187)
 1. [](#bugfix)
     * Fixed CLI progressbar in `backup` and `security` commands to use styled output [#3198](https://github.com/getgrav/grav/issues/3198)
     * Fixed page save failing because of uploaded images [#3191](https://github.com/getgrav/grav/issues/3191)